+++ internal/backend/remote-state/azure/auth/oidc_auth.go
@@ -8,11 +8,9 @@ package auth
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"net/http"
-	"os"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
@@ -38,22 +36,27 @@ func (cred *oidcAuth) Name() string {
 
 func (cred *oidcAuth) Construct(ctx context.Context, config *Config) (azcore.TokenCredential, error) {
 	client := httpclient.New(ctx)
+
+	clientId, err := consolidateClientId(config)
+	if err != nil {
+		// This should never happen; this is checked in the Validate function
+		return nil, err
+	}
 	var token string
-	var err error
 	if config.OIDCToken == "" && config.OIDCTokenFilePath == "" {
 		token, err = getTokenFromRemote(client, config.OIDCAuthConfig)
 		if err != nil {
 			return nil, err
 		}
 	} else {
-		token, err = consolidateToken(config.OIDCAuthConfig)
+		token, err = consolidateToken(config)
 		if err != nil {
 			return nil, err
 		}
 	}
 	return azidentity.NewClientAssertionCredential(
 		config.TenantID,
-		config.ClientID,
+		clientId,
 		func(_ context.Context) (string, error) {
 			return token, nil
 		},
@@ -103,22 +106,8 @@ func getTokenFromRemote(client *http.Client, config OIDCAuthConfig) (string, err
 	return token.Value, nil
 }
 
-func consolidateToken(config OIDCAuthConfig) (string, error) {
-	token := config.OIDCToken
-	if config.OIDCTokenFilePath == "" {
-		return token, nil
-	}
-
-	b, err := os.ReadFile(config.OIDCTokenFilePath)
-	if err != nil {
-		return "", fmt.Errorf("error reading token file: %w", err)
-	}
-
-	fileToken := string(b)
-	if token != "" && token != fileToken {
-		return "", errors.New("token provided directly and through file do not match; either make them the same value or only provide one")
-	}
-	return fileToken, nil
+func consolidateToken(config *Config) (string, error) {
+	return consolidateFileAndValue(config.OIDCToken, config.OIDCTokenFilePath, "token", true)
 }
 
 func (cred *oidcAuth) Validate(ctx context.Context, config *Config) tfdiags.Diagnostics {
@@ -126,23 +115,24 @@ func (cred *oidcAuth) Validate(ctx context.Context, config *Config) tfdiags.Diag
 	if !config.UseOIDC {
 		diags = diags.Append(tfdiags.Sourceless(
 			tfdiags.Error,
-			"Azure OpenID Connect Auth: use_oidc set to false",
-			"use_oidc or the environment variable ARM_USE_OIDC must be set to true",
+			"Invalid Azure OpenID Connect Auth",
+			"OpenID Connect Auth is disabled when use_oidc or the environment variable ARM_USE_OIDC are unset or set explicitly to false.",
 		))
 		return diags
 	}
 	if config.TenantID == "" {
 		diags = diags.Append(tfdiags.Sourceless(
 			tfdiags.Error,
-			"Azure OpenID Connect Auth: missing Tenant ID",
-			"Tenant ID is required",
+			"Invalid Azure OpenID Connect Auth",
+			"Tenant ID is missing.",
 		))
 	}
-	if config.ClientID == "" {
+	_, err := consolidateClientId(config)
+	if err != nil {
 		diags = diags.Append(tfdiags.Sourceless(
 			tfdiags.Error,
-			"Azure OpenID Connect Auth: missing Client ID",
-			"Client ID is required",
+			"Invalid Azure OpenID Connect Auth",
+			fmt.Sprintf("The Client ID is misconfigured: %s.", tfdiags.FormatError(err)),
 		))
 	}
 	directTokenUnset := config.OIDCToken == "" && config.OIDCTokenFilePath == ""
@@ -150,8 +140,8 @@ func (cred *oidcAuth) Validate(ctx context.Context, config *Config) tfdiags.Diag
 	if directTokenUnset && indirectTokenUnset {
 		diags = diags.Append(tfdiags.Sourceless(
 			tfdiags.Error,
-			"Azure OpenID Connect Auth: missing access token",
-			"An access token must be provided, either directly with a variable or through a file, or indirectly through a request URL and request token (as in GitHub Actions)",
+			"Invalid Azure OpenID Connect Auth",
+			"An access token must be provided, either directly with a variable or through a file, or indirectly through a request URL and request token (as in GitHub Actions).",
 		))
 	}
 	if directTokenUnset {
@@ -160,17 +150,17 @@ func (cred *oidcAuth) Validate(ctx context.Context, config *Config) tfdiags.Diag
 		if err != nil {
 			diags = diags.Append(tfdiags.Sourceless(
 				tfdiags.Error,
-				"Azure OpenID Connect Auth: error fetching token",
-				fmt.Sprintf("The following error was encountered while fetching the token: %s", err.Error()),
+				"Invalid Azure OpenID Connect Auth",
+				fmt.Sprintf("Tried to test fetching the token, but received this error instead: %s.", tfdiags.FormatError(err)),
 			))
 		}
 	}
 	// This will work, even if both token and file path are empty
-	if _, err := consolidateToken(config.OIDCAuthConfig); err != nil {
+	if _, err := consolidateToken(config); err != nil {
 		diags = diags.Append(tfdiags.Sourceless(
 			tfdiags.Error,
-			"Azure OpenID Connect Auth: error in token configuration",
-			fmt.Sprintf("The following error was encountered: %s", err.Error()),
+			"Invalid Azure OpenID Connect Auth",
+			fmt.Sprintf("The token is misconfigured: %s", err.Error()),
 		))
 	}
 	return diags